requirements.txt
folium==0.15.0
haversine==2.8.0
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
Pillow==10.1.0
plotly==5.18.0
streamlit==1.28.2
streamlit_folium==0.15.1
